refactor(book): 가격 타입 변경
- BigDecimal에서 Long으로 수정

+++ src/main/java/com/example/ililbooks/domain/book/dto/request/BookCreateRequest.java
@@ -6,8 +6,6 @@ import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-import java.math.BigDecimal;
-
 import static com.example.ililbooks.global.dto.ValidationMessage.*;
 
 @Getter
@@ -21,7 +19,7 @@ public class BookCreateRequest {
     private String author;
 
     @NotNull(message = NOT_NULL_PRICE)
-    private BigDecimal price;
+    private Long price;
 
     private String category;
