+++ src/main/java/com/example/ililbooks/domain/book/dto/request/BookUpdateRequest.java
@@ -1,14 +1,10 @@
 package com.example.ililbooks.domain.book.dto.request;
 
-import com.example.ililbooks.domain.book.enums.LimitedType;
-import com.example.ililbooks.domain.book.enums.SaleStatus;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-
-import java.math.BigDecimal;
 
 import static com.example.ililbooks.global.dto.ValidationMessage.*;
 import static com.example.ililbooks.global.dto.ValidationMessage.INVALID_STOCK;
@@ -23,7 +19,7 @@ public class BookUpdateRequest {
     private String author;
 
     @NotNull(message = NOT_NULL_PRICE)
-    private BigDecimal price;
+    private Long price;
 
     @NotNull(message = NOT_NULL_PRICE)
     private String category;
